Lab6/consumer2/requirements.txt
kafka-python
minio
pandas